Developing a Logout  Feature & debugging my logout feature
I made a log out feature that will then tell the user that they can use
to 1. log out, 2. give a message that the user is not login to the
system anymore.

I had problems with the Session ID not connecting, I had to debug the
problem by checking if all the variables and strings matched. (Which
they did not)